Attempt 1 at catching up the Doubleshot for CM9 glory: This is rebase of the doubleshot repo, based off a KANG of shooteru, combined with previously known good ICS 4.0.3 configuration.
Currently the kernel doesn't boot, so this doesn't work.  But it cleanly
compiles, and once the kernel works, I think it should be usable.

The kernel is from the unified msm8660 kernel found on toastcfh github.
It currently bootloops at kernel load.

I will update as soon as I have a chance to get a good working kernel, or
return to a known good 4.0.3 kernel from December.