doubleshot: Re-enable BT quick switch now that wakelock is fixed
With the catch up of upstream AOSP and Pyramid kernel base this
overlay option is no longer needed to prevent Bluetooth wake_lock
from preventing deep sleep when paired or turned off.

This switch has been disabled in pyramid as of 20120715.  On
Doubleshot, it appears bluetooth properly sleeps with switch
enabled.